requirements.txt
fastmcp>=0.2.0
supabase>=1.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
